Condition == evaluation fix
Added missing condition `== ` and `!=` evaluation for text expressions.
For example, the layout section can now contain the condition
as follows:

`
"condition": "model.name=='abc'"
`

In order to show certain property base on the name property value.

Or the negative one:

`
"condition": "model.name!='abc'"
`

In order to show certain property only whne the property value is not
set or is not equal to `abc`.

Fixes: `Condition with == does not work #98`